requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cloudinary==1.26.0
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.2
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
